+++ src/doc/models/DocResource.ts
