+++ src/doc/validation/ValidationRule.spec.ts
@@ -5,7 +5,7 @@ import {
 	TestFixture,
 } from "alsatian";
 import { TestCase } from "alsatian/core/decorators";
-import IValidationRuleConfig from "../config/IValidationRuleConfig";
+import IValidationRuleConfig from "../../config/IValidationRuleConfig";
 import { ValidatorRule } from "./ValidatorRule";
 
 /* tslint:disable:max-classes-per-file completed-docs */
