+++ src/gm_project/gms1/GMS1Descriptor.d.ts
