+++ src/gm_project/gms1/GMS1Folder.ts
@@ -1,4 +1,4 @@
-import { IGMFolder } from "../IGMInterfaces";
+import IGMFolder from "../interfaces/IGMFolder";
 import GMS1Resource from "./GMS1Resource";
 
 /**
