+++ src/gm_project/gms1/GMS1Resource.spec.ts
@@ -4,10 +4,11 @@ import {
 	TestFixture,
 } from "alsatian";
 
-import { IGMFolder, IGMResource } from "../IGMInterfaces";
-import GMS2Resource from "./GMS2Resource";
+import IGMFolder from "../interfaces/IGMFolder";
+import IGMResource from "../interfaces/IGMResource";
+import GMS1Resource from "./GMS1Resource";
 
-/* tslint:disable:max-classes-per-file completed-docs */
+/* tslint:disable:completed-docs max-classes-per-file */
 
 class FolderMock implements IGMFolder {
 	public parent: IGMFolder | null;
@@ -16,14 +17,9 @@ class FolderMock implements IGMFolder {
 	public children: IGMResource[];
 }
 
-@TestFixture("GMS2Resource")
-export class GMS2ResourceFixture {
-
-	public resource = new GMS2Resource({
-		// IResource
-		id: "my-id",
-		name: "my-name",
-	});
+@TestFixture("GMS1Resource")
+export class GMS1ResourceFixture {
+	public resource = new GMS1Resource(null, "my-name");
 
 	@Test("should get the name")
 	public name() {
@@ -31,12 +27,12 @@ export class GMS2ResourceFixture {
 	}
 
 	@Test("should get the fullpath when it has no parent")
-	public fullpath_NoParent() {
+	public fullpath() {
 		Expect(this.resource.fullpath).toBe("my-name");
 	}
 
 	@Test("should get the fullpath when it has parent")
-	public fullpath_Parent() {
+	public fullpath_HasParent() {
 		const folder = new FolderMock();
 		folder.fullpath = "my-fullpath/";
 		this.resource.parent = folder;
