+++ src/gm_project/gms1/GMS1Script.spec.ts
