+++ src/gm_project/gms2/GMS2ResourceType.ts
