+++ src/gm_project/gms2/GMS2Script.spec.ts
