+++ src/gm_project/gms2/IGMS2Descriptor.d.ts
