+++ src/gm_project/gms2/IGetResourceByKey.d.ts
@@ -1,4 +1,4 @@
-import { IGMResource } from '../IGMInterfaces';
+import IGMResource from '../interfaces/IGMResource';
 
 export default interface IGetResourceByKey {
 	getResourceByKey(key: string): IGMResource | undefined;
