+++ src/gm_project/interfaces/IGMFolder.d.ts
@@ -1,0 +1,8 @@
+import IGMResource from "./IGMResource";
+
+/**
+ * Represents a Single GameMaker Folder
+ */
+export default interface IGMFolder extends IGMResource {
+	readonly children: IGMResource[];
+}
