+++ src/gm_project/interfaces/IGMProject.d.ts
@@ -1,0 +1,10 @@
+import IGMResource from "./IGMResource";
+
+/**
+ * Represents a single GameMaker Project
+ */
+export default interface IGMProject {
+	readonly path: string;
+	readonly name: string;
+	find(pattern: string): IGMResource[];
+}
