+++ src/gm_project/interfaces/IGMResource.d.ts
@@ -1,0 +1,10 @@
+import IGMFolder from "./IGMFolder";
+
+/**
+ * Represents a single GameMaker Resource
+ */
+export default interface IGMResource {
+	parent: IGMFolder | null;
+	fullpath: string;
+	name: string;
+}
